Created Test and License Backend

+++ DVLD_DataAccess/Models/ApplicationTypeModel.cs
@@ -6,9 +6,9 @@ namespace DVLD_DataAccess.Models
     {
         public int Id { set; get; }
         public required string Title { set; get; }
-        public decimal Fees { set; get; }
+        public float Fees { set; get; }
         [SetsRequiredMembers]
-        public ApplicationTypeModel(int id, string title, decimal fees)
+        public ApplicationTypeModel(int id, string title, float fees)
         {
             Id = id;
             Title = title;
